No new code, just moving files around to newly created test directories for much better organization.
