+++ app/src/androidTest/java/com/heyzeusv/solitaire/di/TestAppModule.kt
@@ -1,10 +1,10 @@
-package com.heyzeusv.solitaire
+package com.heyzeusv.solitaire.di
 
 import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.core.DataStoreFactory
 import androidx.datastore.dataStoreFile
-import com.heyzeusv.solitaire.di.AppModule
+import com.heyzeusv.solitaire.StatPreferences
 import com.heyzeusv.solitaire.util.StatPreferencesSerializer
 import dagger.Module
 import dagger.Provides
