+++ app/src/androidTest/java/com/heyzeusv/solitaire/ui/PileTest.kt
@@ -1,4 +1,4 @@
-package com.heyzeusv.solitaire
+package com.heyzeusv.solitaire.ui
 
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.test.assertHasClickAction
@@ -9,8 +9,10 @@ import androidx.compose.ui.test.isDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.unit.dp
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.heyzeusv.solitaire.ui.SolitairePile
+import com.heyzeusv.solitaire.R
 import com.heyzeusv.solitaire.ui.theme.SolitaireTheme
+import com.heyzeusv.solitaire.util.TestCards
+import com.heyzeusv.solitaire.util.onNodeWithConDescId
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
