+++ app/src/androidTest/java/com/heyzeusv/solitaire/ui/ScoreboardTest.kt
@@ -1,4 +1,4 @@
-package com.heyzeusv.solitaire
+package com.heyzeusv.solitaire.ui
 
 import androidx.activity.ComponentActivity
 import androidx.compose.foundation.layout.height
@@ -7,8 +7,9 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.unit.dp
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.heyzeusv.solitaire.ui.SolitaireScoreboard
+import com.heyzeusv.solitaire.R
 import com.heyzeusv.solitaire.util.SolitairePreview
+import com.heyzeusv.solitaire.util.onNodeWithTextId
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
