+++ app/src/test/java/com/heyzeusv/solitaire/ui/TableauTest.kt
@@ -1,7 +1,8 @@
-package com.heyzeusv.solitaire
+package com.heyzeusv.solitaire.ui
 
 import com.heyzeusv.solitaire.data.Card
 import com.heyzeusv.solitaire.data.Tableau
+import com.heyzeusv.solitaire.util.TestCards
 import org.junit.Assert.assertEquals
 import org.junit.Test
